+++ ambassador/ambassador/envoy/v2/v2listener.py
@@ -12,14 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
-from typing import List, Optional, TYPE_CHECKING
+from typing import Any, Dict, List, Optional, TYPE_CHECKING
 from typing import cast as typecast
 
 from copy import deepcopy
 
 from multi import multi
 from ...ir.irlistener import IRListener
+from ...ir.irauth import IRAuth
 from ...ir.irfilter import IRFilter
+from ...ir.irratelimit import IRRateLimit
+from ...ir.ircors import IRCORS
+from ...ir.ircluster import IRCluster
 
 from .v2tls import V2TLSContext
 from .v2route import V2Route
@@ -27,7 +31,28 @@ from .v2route import V2Route
 if TYPE_CHECKING:
     from . import V2Config
 
+# Static header keys normally used in the context of and authorization request.
+AllowedRequestHeaders = frozenset([
+    'authorization',
+    'cookie',
+    'from',
+    'proxy-authorization',
+    'user-agent',
+    'x-forwarded-for',
+    'x-forwarded-host',
+    'x-forwarded-proto'
+])
 
+# Static header keys normally used in the context of and authorization response.
+AllowedAuthorizationHeaders = frozenset([
+    'location',
+    'authorization',
+    'proxy-authenticate',
+    'set-cookie',
+    'www-authenticate'
+])
+
+# This mapping is only used for ambassador/v0.
 ExtAuthRequestHeaders = {
     'Authorization': True,
     'Cookie': True,
@@ -38,23 +63,45 @@ ExtAuthRequestHeaders = {
     'Proxy-Authorization': True,
     'Set-Cookie': True,
     'User-Agent': True,
+    'x-b3-flags': True,
+    'x-b3-parentspanid': True,
+    'x-b3-traceid': True,
+    'x-b3-sampled': True,
+    'x-b3-spanid': True,
     'X-Forwarded-For': True,
     'X-Forwarded-Host': True,
     'X-Forwarded-Proto'
     'X-Gateway-Proto': True,
+    'x-ot-span-context': True,
     'WWW-Authenticate': True,
 }
 
 @multi
-def v2filter(irfilter):
+def v2filter(irfilter: IRFilter):
     return irfilter.kind
 
-@v2filter.when("IRAuth")
-def v2filter(auth):
-    request_headers = dict(ExtAuthRequestHeaders)
 
-    for hdr in auth.allowed_headers:
-        request_headers[hdr] = True
+@v2filter.when("IRAuth")
+def v2filter_auth(auth: IRAuth):
+    if auth.api_version == "ambassador/v0":
+        # This preserves exactly the same logic prior to ambassador/v1 implementation.
+        request_headers = dict(ExtAuthRequestHeaders)
+
+        for hdr in auth.allowed_headers:
+            request_headers[hdr] = True
+
+        allowed_authorizarion_headers = auth.allowed_headers
+        allowed_request_headers = sorted(request_headers.keys())
+    else:
+        # We only really know how to handle v1 or v0. Hmm.
+        assert auth.api_version == "ambassador/v1"
+
+        allowed_authorizarion_headers = list(set(auth.allowed_authorization_headers).union(AllowedAuthorizationHeaders))
+        allowed_request_headers = list(set(auth.allowed_request_headers).union(AllowedRequestHeaders))
+    
+    assert auth.cluster
+
+    cluster = typecast(IRCluster, auth.cluster)
 
     return {
         'name': 'envoy.ext_authz',
@@ -62,20 +109,19 @@ def v2filter(auth):
             'http_service': {
                 'server_uri': {
                     'uri': 'http://%s' % auth.auth_service,
-                    'cluster': auth.cluster.name,
-                    'timeout': '3s',
+                    'cluster': cluster.name,
+                    'timeout': "%0.3fs" % (float(auth.timeout_ms) / 1000.0)
                 },
                 'path_prefix': auth.path_prefix,
-                'allowed_authorization_headers': auth.allowed_headers,
-                'allowed_request_headers': sorted(request_headers.keys())
-                # 'authorization_headers_to_add': []
+                'allowed_authorization_headers': allowed_authorizarion_headers,
+                'allowed_request_headers': allowed_request_headers
             }
-        }
-    }
+        }        
+    }    
 
 
 @v2filter.when("IRRateLimit")
-def v2filter(ratelimit):
+def v2filter_ratelimit(ratelimit: IRRateLimit):
     config = dict(ratelimit.config)
 
     if 'timeout_ms' in config:
@@ -88,13 +134,17 @@ def v2filter(ratelimit):
         'config': config
     }
 
+
 @v2filter.when("ir.cors")
-def v2filter(cors):
+def v2filter_cors(cors: IRCORS):
+    del cors    # silence unused-variable warning
+
     return { 'name': 'envoy.cors' }
 
+
 @v2filter.when("ir.router")
-def v2filter(router):
-    od = { 'name': 'envoy.router' }
+def v2filter_router(router: IRFilter):
+    od: Dict[str, Any] = { 'name': 'envoy.router' }
 
     if router.ir.tracing:
         od['config'] = { 'start_child_span': True }
@@ -106,17 +156,13 @@ class V2Listener(dict):
     def __init__(self, config: 'V2Config', listener: IRListener) -> None:
         super().__init__()
 
-        # TODO: rate limit
-        # if "rate_limits" in group:
-        #     route["rate_limits"] = group.rate_limits
-
         # Default some things to the way they should be for the redirect listener
         name = "redirect_listener"
         envoy_ctx: Optional[dict] = None
         access_log: Optional[List[dict]] = None
         require_tls: Optional[str] = 'ALL'
         use_proxy_proto: Optional[bool] = None
-        filters: List[dict] = [ { 'name': 'envoy.router' } ]
+        filters: List[dict] = []
         routes: List[V2Route] = typecast(List[V2Route], [ {
             'match': {
                 'prefix': '/',
@@ -152,9 +198,9 @@ class V2Listener(dict):
             config.ir.logger.info("envoy_ctx final %s" % envoy_ctx)
 
             # Assemble filters
-            filters = []
             for f in config.ir.filters:
-                v2f = v2filter(f)
+                v2f: dict = v2filter(f)
+
                 if v2f:
                     filters.append(v2f)
 
@@ -177,46 +223,42 @@ class V2Listener(dict):
         if require_tls:
             vhost['require_tls'] = require_tls
 
-        hcm_config = {
-                'name': 'envoy.http_connection_manager',
-                'config': {
-                    'stat_prefix': 'ingress_http',
-                    'access_log': access_log,
-                    'http_filters': filters,
-                    'route_config': {
-                        'virtual_hosts': [ vhost ]
-                    }
-                }
+        http_config: Dict[str, Any] = {
+            'stat_prefix': 'ingress_http',
+            'access_log': access_log,
+            'http_filters': filters,
+            'route_config': {
+                'virtual_hosts': [ vhost ]
             }
+        }
 
         for group in config.ir.ordered_groups():
             if group.get('use_websocket'):
-                hcm_config['config'].update(
-                    {
-                        'upgrade_configs': [
-                            {
-                                'upgrade_type': 'websocket',
-                            },
-                        ]
-                    },
-                )
+                http_config['upgrade_configs'] = [ { 'upgrade_type': 'websocket' } ]
                 break
 
-        hcm_config_conf = hcm_config["config"]
+        if 'use_remote_address' in config.ir.ambassador_module:
+            http_config["use_remote_address"] = config.ir.ambassador_module.use_remote_address
 
         if config.ir.tracing:
-            hcm_config_conf["generate_request_id"] = True
-            hcm_config_conf["tracing"] = {
+            http_config["generate_request_id"] = True
+
+            http_config["tracing"] = {
                 "operation_name": "egress",
             }
 
             req_hdrs = config.ir.tracing.get('tag_headers', [])
 
             if req_hdrs:
-                hcm_config_conf["tracing"]["request_headers_for_tags"] = req_hdrs
+                http_config["tracing"]["request_headers_for_tags"] = req_hdrs
 
-        chain = {
-            'filters': [hcm_config]
+        http_connmgr_config = {
+                'name': 'envoy.http_connection_manager',
+                'config': http_config
+            }
+
+        chain: Dict[str, Any] = {
+            'filters': [ http_connmgr_config ]
         }
 
         if envoy_ctx:   # envoy_ctx has to exist _and_ not be empty to be truthy
@@ -249,8 +291,15 @@ class V2Listener(dict):
 
     def handle_sni(self, config: 'V2Config'):
         global_sni = False
-        # Pretty sure we will have just one chain
-        filter_chain = self.get('filter_chains')[0]
+        filter_chains = self.get('filter_chains', [])
+
+        # We really must have a filter chain.
+        assert filter_chains
+
+        # We really should have just one chain.
+        assert len(filter_chains) == 1
+
+        filter_chain = filter_chains[0]
 
         for tls_context in config.ir.tls_contexts:
             if not global_sni:
@@ -303,7 +352,6 @@ class V2Listener(dict):
                     'private_key_file']
 
                 if hosts_match and certificate_chain_file_match and private_key_file_match:
-
                     chain['filters'][0]['config']['route_config']['virtual_hosts'][0]['routes'].append(sni_route['route'])
 
             self['filter_chains'].append(chain)