+++ ambassador/ambassador/ir/irauth.py
@@ -1,4 +1,5 @@
 from typing import Optional, TYPE_CHECKING
+from typing import cast as typecast
 
 from ..config import Config
 from ..utils import RichStatus
@@ -12,20 +13,25 @@ if TYPE_CHECKING:
 
 
 class IRAuth (IRFilter):
+    cluster: Optional[IRCluster]
+
     def __init__(self, ir: 'IR', aconf: Config,
                  rkey: str="ir.auth",
                  kind: str="IRAuth",
                  name: str="extauth",
                  type: Optional[str] = "decoder",
                  **kwargs) -> None:
-        # print("IRAuth __init__ (%s %s %s)" % (kind, name, kwargs))
+
 
         super().__init__(
             ir=ir, aconf=aconf, rkey=rkey, kind=kind, name=name,
-            cluster="cluster_ext_auth",
+            cluster=None,
             timeout_ms=5000,
             path_prefix=None,
+            api_version=None,
             allowed_headers=[],
+            allowed_request_headers=[],
+            allowed_authorization_headers=[],
             hosts={},
             type=type,
             **kwargs)
@@ -39,7 +45,6 @@ class IRAuth (IRFilter):
         config_info = aconf.get_config("auth_configs")
 
         if config_info:
-            # self.logger.debug("auth_configs: %s" % auth_configs)
             for config in config_info.values():
                 self._load_auth(config)
 
@@ -52,11 +57,10 @@ class IRAuth (IRFilter):
         return True
 
     def add_mappings(self, ir: 'IR', aconf: Config):
-        cluster_hosts = self.get('hosts', { '127.0.0.1:5000': ( 100, None ) })
+        cluster_hosts = self.get('hosts', { '127.0.0.1:5000': ( 100, None, '-internal-' ) })
 
         self.cluster = None
-
-        # self.ir.logger.debug("AUTH ADD_MAPPINGS: %s" % self.as_json())
+        cluster_good = False
 
         for service, params in cluster_hosts.items():
             weight, ctx_name, location = params
@@ -67,8 +71,6 @@ class IRAuth (IRFilter):
                 host_rewrite=self.get('host_rewrite', False),
                 ctx_name=ctx_name,
                 marker='extauth'
-                # grpc=self.get('grpc', False)
-
             )
 
             cluster.referenced_by(self)
@@ -79,46 +81,13 @@ class IRAuth (IRFilter):
                 if not self.cluster.merge(cluster):
                     self.post_error(RichStatus.fromError("auth canary %s can only change service!" % cluster.name))
                     cluster_good = False
-                    # self.ir.logger.debug("BAD MERGE %s" % cluster.as_json())
-                # else:
-                #     self.ir.logger.debug("GOOD MERGE %s" % cluster.as_json())
             else:
                 self.cluster = cluster
-                # self.ir.logger.debug("SET CLUSTER %s" % cluster.as_json())
 
         if cluster_good:
-            # self.ir.logger.debug("GOOD CLUSTER %s" % self.cluster.as_json())
-            ir.add_cluster(self.cluster)
-            self.referenced_by(self.cluster)
+            ir.add_cluster(typecast(IRCluster, self.cluster))
+            self.referenced_by(typecast(IRCluster, self.cluster))
 
-        #     urls = []
-        #     protocols = {}
-        #
-        #     for svc in sorted(cluster_hosts.keys()):
-        #         _, tls_context = cluster_hosts[svc]
-        #
-        #         (svc, url, originate_tls, otls_name) = self.service_tls_check(svc, tls_context, host_rewrite)
-        #
-        #         if originate_tls:
-        #             protocols['https'] = True
-        #         else:
-        #             protocols['http'] = True
-        #
-        #         if otls_name:
-        #             filter_config['cluster'] = cluster_name + "_" + otls_name
-        #             cluster_name = filter_config['cluster']
-        #
-        #         urls.append(url)
-        #
-        #     if len(protocols.keys()) != 1:
-        #         raise Exception("auth config cannot try to use both HTTP and HTTPS")
-        #
-        #     self.add_intermediate_cluster(first_source, cluster_name,
-        #                                   'extauth', urls,
-        #                                   type="strict_dns", lb_type="round_robin",
-        #                                   originate_tls=originate_tls, host_rewrite=host_rewrite)
-        #
-        # name = "internal_%s_probe_mapping" % name
 
     def _load_auth(self, module: Resource):
         if self.location == '--internal--':
@@ -142,24 +111,24 @@ class IRAuth (IRFilter):
 
             self.referenced_by(module)
 
-        headers = module.get('allowed_headers', None)
-
-        if headers:
-            allowed_headers = self.get('allowed_headers', [])
-
-            for hdr in headers:
-                if hdr not in allowed_headers:
-                    allowed_headers.append(hdr)
-
-            self['allowed_headers'] = allowed_headers
+        self["api_version"] = module.get("apiVersion", None)
+        self["timeout_ms"] = module.get("timeout_ms", 5000)
+        
+        self.__to_header_list('allowed_headers', module)
+        self.__to_header_list('allowed_request_headers', module)
+        self.__to_header_list('allowed_authorization_headers', module)
 
         auth_service = module.get("auth_service", None)
-        # weight = module.get("weight", 100)
         weight = 100    # Can't support arbitrary weights right now.
 
         if auth_service:
             self.hosts[auth_service] = ( weight, module.get('tls', None), module.location )
 
+        # IRAuth requires this in order to support ambassador/v0 and ambassador/v1 version. 
+        if self["api_version"] == None:
+            self.post_error(RichStatus.fromError("Missing apiVersion in Auth configuration"))
+
+    # This method is only used by v1listener.
     def config_dict(self):
         config = {
             "cluster": self.cluster.name
@@ -173,3 +142,15 @@ class IRAuth (IRFilter):
             config['allowed_headers'] = self.allowed_headers
 
         return config
+
+    def __to_header_list(self, list_name, module):
+        headers = module.get(list_name, None)
+
+        if headers:
+            allowed_headers = self.get(list_name, [])
+        
+            for hdr in headers:
+                if hdr not in allowed_headers:
+                    allowed_headers.append(hdr.lower())
+
+            self[list_name] = allowed_headers