+++ ambassador/tests/kat/t_ratelimit.py
@@ -1,17 +1,11 @@
-import json
-import pytest
+from kat.harness import Query
 
-from typing import ClassVar, Dict, List, Sequence, Tuple, Union
-
-from kat.harness import sanitize, variants, Query, Runner
-from kat import manifests
-
-from abstract_tests import AmbassadorTest, HTTP
-from abstract_tests import MappingTest, OptionTest, ServiceType, Node, Test
+from abstract_tests import AmbassadorTest, HTTP, ServiceType
 
 
 class RateLimitTest(AmbassadorTest):
     # debug = True
+    target: ServiceType
 
     def init(self):
         self.target = HTTP()